# [1/3] Wrap simdjson kernels in anonymous namespaces