--- a/src/arm64/bitmanipulation.h
+++ b/src/arm64/bitmanipulation.h
@@ -55,6 +55,7 @@ really_inline bool add_overflow(uint64_t value1, uint64_t value2, uint64_t *resu
 #endif
 }
 
+#if 0 // Currently unused
 really_inline bool mul_overflow(uint64_t value1, uint64_t value2, uint64_t *result) {
 #ifdef SIMDJSON_REGULAR_VISUAL_STUDIO
   *result = value1 * value2;
@@ -63,6 +64,7 @@ really_inline bool mul_overflow(uint64_t value1, uint64_t value2, uint64_t *resu
   return __builtin_umulll_overflow(value1, value2, (unsigned long long *)result);
 #endif
 }
+#endif // Currently unused
 
 } // namespace arm64
 } // namespace {