--- a/src/arm64/bitmask.h
+++ b/src/arm64/bitmask.h
@@ -1,9 +1,8 @@
 #ifndef SIMDJSON_ARM64_BITMASK_H
 #define SIMDJSON_ARM64_BITMASK_H
 
-namespace simdjson {
-namespace arm64 {
 namespace {
+namespace arm64 {
 
 //
 // Perform a "cumulative bitwise xor," flipping bits each time a 1 is encountered.