--- a/src/generic/stage1/allocate.h
+++ b/src/generic/stage1/allocate.h
@@ -1,5 +1,6 @@
 namespace simdjson {
 namespace SIMDJSON_IMPLEMENTATION {
+namespace {
 namespace stage1 {
 namespace allocate {
 
@@ -17,5 +18,6 @@ really_inline error_code set_capacity(internal::dom_parser_implementation &parse
 
 } // namespace allocate
 } // namespace stage1
+} // namespace {
 } // namespace SIMDJSON_IMPLEMENTATION
 } // namespace simdjson