First in a three-part PR series: 1065 (anonymous namespaces) -- #1078 (remove computed GOTOs) -- #1079 (return errors instead of goto).

This:

* Moves simdjson::haswell, simdjson::arm64, simdjson::westmere and simdjson::fallback to the top level namespace (outside of simdjson, which is for public stuff). This makes it clear these are not for public consumption as part of the main package.
* Wraps these namespaces in anonymous namespaces, giving the compiler license to apply any private optimizations it wishes. It might also prevent them from being exported, which I think we have an issue for somewhere.

The goal here is to not expose these things publicly, and thereby give the compiler more license to optimize them.

## Performance

On Intel gcc9, this is a loss, which appears to be due to jitter. Instruction count is generally up. I suspect binary layout is the key problem with our Intel jitter, based on the fact that the numbers are reversing an unexplainable *gain* in performance found in [this PR](https://github.com/simdjson/simdjson/pull/1082).

On ARM gcc9, it's a very minor win, reducing instruction count more reliably.

### skylake-x gcc9

| File                             | Blocks | master Cycles | PR Cycles | +Throughput | master Instr. | PR Instr. | -Instr. |
| ---                              |    --: |   --: |   --: |   --: |   --: |   --: |   --: |
| citm_catalog                     |  26987 |  84.5 |  83.7 |    0% | 312.6 | 312.5 |    0% |
| gsoc-2018                        |  51997 |  73.5 |  73.2 |    0% | 168.5 | 169.0 |    0% |
| github_events                    |   1017 |  77.9 |  77.6 |    0% | 266.5 | 267.3 |    0% |
| numbers                          |   2345 | 241.9 | 241.1 |    0% | 786.6 | 799.3 |   -1% |
| twitter                          |   9867 |  92.1 |  91.9 |    0% | 295.7 | 296.9 |    0% |
| update-center                    |   8330 | 112.7 | 112.8 |    0% | 340.4 | 342.6 |    0% |
| canada                           |  35172 | 294.8 | 295.7 |    0% | 985.1 | 982.0 |    0% |
| twitterescaped                   |   8787 | 189.8 | 191.3 |    0% | 507.0 | 512.0 |    0% |
| marine_ik                        |  46616 | 288.8 | 291.4 |    0% | 912.0 | 915.6 |    0% |
| apache_builds                    |   1988 |  83.4 |  84.3 |   -1% | 306.0 | 307.4 |    0% |
| random                           |   7976 | 139.7 | 141.6 |   -1% | 496.7 | 496.8 |    0% |
| mesh                             |  11306 | 298.1 | 304.6 |   -2% | 988.2 | 989.7 |    0% |
| mesh.pretty                      |  24646 | 180.7 | 185.0 |   -2% | 583.5 | 584.1 |    0% |
| instruments                      |   3442 | 102.7 | 108.5 |   -5% | 382.1 | 376.1 |    1% |

### ampere gcc9

| File                             | Blocks | master Cycles | PR Cycles | +Throughput | master Instr. | PR Instr. | -Instr. | master Branch Misses | PR Branch Misses | -Branch Misses | master Cache Misses | PR Cache Misses | -Cache Misses | master Cache Refs | PR Cache Refs | -Cache Refs |
| ---                              |    --: |   --: |   --: |   --: |   --: |   --: |   --: |    --: |    --: |   --: |    --: |    --: |   --: |    --: |    --: |   --: |
| citm_catalog                     |  26987 | 399.6 | 392.9 |    1% | 374.4 | 381.8 |   -1% |  11024 |   8585 |   22% |  62741 |  62713 |    0% | 1952696 | 1885644 |    3% |
| random                           |   7976 | 708.0 | 696.8 |    1% | 639.5 | 639.2 |    0% |   6877 |   7602 |  -10% |  21605 |  21699 |    0% | 1222285 | 1207980 |    1% |
| github_events                    |   1017 | 463.9 | 457.1 |    1% | 359.0 | 360.1 |    0% |   1591 |   1402 |   11% |   2411 |   2407 |    0% |  88871 |  86889 |    2% |
| instruments                      |   3442 | 484.9 | 479.1 |    1% | 464.7 | 462.5 |    0% |   3733 |   3325 |   10% |   8685 |   8674 |    0% | 341272 | 333461 |    2% |
| twitterescaped                   |   8787 | 743.0 | 735.0 |    1% | 622.3 | 623.4 |    0% |  18547 |  19381 |   -4% |  21403 |  21360 |    0% | 1575900 | 1563878 |    0% |
| marine_ik                        |  46616 | 897.5 | 887.8 |    1% | 987.8 | 985.3 |    0% |  97901 | 103790 |   -6% | 133870 | 133932 |    0% | 7457789 | 7637591 |   -2% |
| numbers                          |   2345 | 913.2 | 904.3 |    0% | 874.0 | 869.7 |    0% |   3253 |   3274 |    0% |   6022 |   6014 |    0% | 239409 | 260123 |   -8% |
| canada                           |  35172 | 847.0 | 840.0 |    0% | 1015.4 | 1023.5 |    0% |  53035 |  56329 |   -6% |  91698 |  91694 |    0% | 4907920 | 5075437 |   -3% |
| mesh                             |  11306 | 951.4 | 944.9 |    0% | 1062.5 | 1055.2 |    0% |  21099 |  20357 |    3% |  32326 |  32332 |    0% | 1621629 | 1683986 |   -3% |
| mesh.pretty                      |  24646 | 594.2 | 590.6 |    0% | 634.2 | 630.8 |    0% |  40340 |  39803 |    1% |  59060 |  59100 |    0% | 1971579 | 2036904 |   -3% |
| update-center                    |   8330 | 544.8 | 543.2 |    0% | 462.7 | 463.8 |    0% |  14055 |  13706 |    2% |  20820 |  20822 |    0% | 1036300 | 1018939 |    1% |
| gsoc-2018                        |  51997 | 359.9 | 359.3 |    0% | 238.2 | 238.4 |    0% |  39291 |  39345 |    0% | 109625 | 109474 |    0% | 2761943 | 2745053 |    0% |
| twitter                          |   9867 | 488.9 | 488.4 |    0% | 386.7 | 387.7 |    0% |  16054 |  16363 |   -1% |  23434 |  23450 |    0% | 864371 | 851627 |    1% |
| apache_builds                    |   1988 | 470.9 | 472.4 |    0% | 414.2 | 414.7 |    0% |    639 |    688 |   -7% |   4829 |   4836 |    0% | 201650 | 200021 |    0% |