@jkeiser 2020-07-25
This revealed some unused functions in the fallback kernel; I've fixed that.