@jkeiser 2020-07-28
Anecdotally, this may solve the stage 2 jitters. Several other changes I've made that are jittery on master are much more stable when rebased against this patch.

Still has a few fixes to go for a couple of compilers.