@jkeiser 2020-07-28
Everything is fixed now. There are VS16 tests waiting to run still; I'm not sure why that is, but am not particularly worried either.