@jkeiser 2020-08-02
After removing the dead code and rebasing against master, this is now a CLEAR win. I don't know why. Numbers updated ...