@jkeiser 2020-08-02
I'm going to try and figure out why this is before checking in. At the very least it'll give us some idea of where jitters can come from. Either removing the dead code fixed it, in which case we know code size and probably locations in the binary are a factor here; or something in master got worse (and nothing in master should have gotten worse).