@jkeiser 2020-08-02
@lemire The [PR to remove global allocation](https://github.com/simdjson/simdjson/pull/1082) improved g++ performance and instruction count massively:

| File                             | Blocks | master Cycles | PR Cycles | +Throughput | master Instr. | PR Instr. | -Instr. |
| ---                              |    --: |   --: |   --: |   --: |   --: |   --: |   --: |
| instruments                      |   3442 | 110.5 | 102.7 |    7% | 382.1 | 382.1 |    0% |
| random                           |   7976 | 146.9 | 139.8 |    4% | 496.7 | 496.7 |    0% |
| canada                           |  35172 | 305.4 | 292.9 |    4% | 985.1 | 985.1 |    0% |
| citm_catalog                     |  26987 |  87.6 |  84.4 |    3% | 312.6 | 312.6 |    0% |
| apache_builds                    |   1988 |  86.7 |  83.7 |    3% | 306.0 | 306.0 |    0% |
| mesh.pretty                      |  24646 | 186.2 | 180.2 |    3% | 583.5 | 583.5 |    0% |
| mesh                             |  11306 | 306.2 | 296.9 |    3% | 988.2 | 988.2 |    0% |
| update-center                    |   8330 | 116.5 | 113.2 |    2% | 340.4 | 340.4 |    0% |
| github_events                    |   1017 |  79.5 |  77.6 |    2% | 266.5 | 266.5 |    0% |
| marine_ik                        |  46616 | 295.2 | 289.4 |    1% | 912.0 | 912.0 |    0% |
| numbers                          |   2345 | 245.6 | 241.3 |    1% | 786.6 | 786.6 |    0% |
| twitter                          |   9867 |  93.3 |  92.0 |    1% | 295.7 | 295.7 |    0% |
| gsoc-2018                        |  51997 |  73.0 |  73.6 |    0% | 168.5 | 168.5 |    0% |
| twitterescaped                   |   8787 | 190.5 | 192.9 |   -1% | 507.0 | 507.0 |    0% |

I am not 100% sure why.