@lemire 2020-08-02
I am also getting that this PR is a net win.